Backend/requirements.txt
flask
flask-cors
python-dotenv
PyPDF2
python-docx
python-pptx
langchain
chromadb
transformers
sentence-transformers
requests

